Parse-html-filter-select-nodes plugin stable release.
I implemented and tested a plugin that works adequately for both extracting specific
information from an HTML document as well as for excluding certain html tags so that
certain content is not indexed and certain outlinks are ignored.
In a few words this plugin selects nodes and/or excludes nodes to extend the functionality
of nutch to allow for focused crawls.
I tested this plugin several times.